@@ -5,17 +5,16 @@ import Navbar from "./components/Navbar";
 import Skills from "./components/Skills";
 import Work from "./components/Work";
 
-
 function App() {
   return (
-    <div>
+    <>
       <Navbar />
       <Home />
       <About />
       <Skills />
       <Work />
       <Contact />
-    </div>
+    </>
   );
 }
 
